requirements.txt
requests
python-telegram-bot
PyYAML
loguru
